requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
marshmallow==3.20.2
gunicorn==21.2.0
pytest==8.0.2 